Improve SSDS3 accuracy & images

+++ pcukdownload.py
@@ -127,7 +127,14 @@ def inspect_game(name, alt_names):
     r = requests.get(f'http://www.pcengine.co.uk/HTML_Games/{name}.htm')
     soup = BeautifulSoup(r.text)
     title = soup.find('span', attrs={'class': 'title'}).text
-    game["name"] = str.strip(title)
+    names = [str.strip(x) for x in title.split('\n')]
+    game['name'] = " ".join(names)
+
+    alternative_names = []
+
+    if len(names) > 1:
+        for n in names:
+            alternative_names.append(n)
 
     try:
         alternative_names_cand = []
@@ -135,8 +142,6 @@ def inspect_game(name, alt_names):
             span = span.text.split('\n')
             alternative_names_cand = alternative_names_cand + [str.strip(x) for x in span]
             alternative_names_cand = list(filter(lambda x: x != "", alternative_names_cand))
-
-        alternative_names = []
 
         for d in alternative_names_cand+alt_names:
             if "RELEASE DATE:" in d:
@@ -165,9 +170,18 @@ def inspect_game(name, alt_names):
                 if extra is not None:
                     extra = extra.group(1)
                     d = d.replace(f"[{extra}]", "")
-                    alternative_names.append(str.strip(extra))
+                    d = str.strip(extra)
+
+                    alternative_names.append(d)
+                    for dd in extra.split(', '):
+                        if dd.lower() != 'the' and dd not in alternative_names:
+                            alternative_names.append(dd)
+
                 d = str.strip(d)
                 alternative_names.append(d)
+                for dd in d.split(', '):
+                    if dd.lower() != 'the' and dd not in alternative_names:
+                        alternative_names.append(dd)
 
         alternative_names = list(filter(lambda x: x != "", alternative_names))
         if len(alternative_names) > 0:
@@ -194,20 +208,29 @@ def inspect_game(name, alt_names):
     except Exception:
         pass
 
-    db.append(game)
 
     # Download game cover
+    if "YAMAMURA MISA SUSPENSE" in game["name"]:
+        game["name"] = game["name"].replace(":", "")
     table = soup.find('table', id="Info and Cover")
     imgs = table.findChildren('img')
     for img in imgs:
         if "Images-Coverthumbs" in img['src']:
-            r = requests.get(f"http://www.pcengine.co.uk/{img['src'][3:]}")
-            with open(f'output/pce/{game["name"].replace("/", " ")}.jpg', 'wb') as f:
-                f.write(r.content)
-            call(f'convert "output/pce/{game["name"].replace("/", " ")}.jpg" -filter lanczos -resize 64x64! -normalize -unsharp 0 -enhance +dither -colors 64 "output/pce/{game["name"].replace("/", " ")}.png"', shell=True)
-            optimize_image(f'output/pce/{game["name"].replace("/", " ")}.png')
-            return (True, game["name"].replace("/", " "))
-    return (False, game["name"].replace("/", " "))
+            if not os.path.isfile(f'output/pce/{game["name"].replace("/", " ")}.jpg'):
+                r = requests.get(f"http://www.pcengine.co.uk/{img['src'][3:]}")
+                with open(f'output/pce/{game["name"].replace("/", " ")}.jpg', 'wb') as f:
+                    f.write(r.content)
+                call(f'convert "output/pce/{game["name"].replace("/", " ")}.jpg" -filter lanczos -resize 64x64! -normalize -unsharp 0 -enhance +dither "output/pce/{game["name"].replace("/", " ")}.bmp3"', shell=True)
+                os.rename(f'output/pce/{game["name"].replace("/", " ")}.bmp3', f'output/pce/{game["name"].replace("/", " ")}.bmp')
+                call(f'python3 Quantomatic.py "output/pce/{game["name"].replace("/", " ")}" -p 13', shell=True)
+                call(f'convert "output/pce/{game["name"].replace("/", " ")}CrushedPals.bmp" "output/pce/{game["name"].replace("/", " ")}.png"', shell=True)
+                os.remove(f'output/pce/{game["name"].replace("/", " ")}.bmp')
+                os.remove(f'output/pce/{game["name"].replace("/", " ")}CrushedPals.bmp')
+            #optimize_image(f'output/pce/{game["name"].replace("/", " ")}.png')
+
+            game['image'] = f'output/pce/{game["name"].replace("/", " ")}.png'
+            return (True, game["name"].replace("/", " "), game)
+    return (False, game["name"].replace("/", " "), game)
 
 
 for c in string.ascii_uppercase:
@@ -223,6 +246,7 @@ for c in string.ascii_uppercase:
         alt_names = link.text.split('\n')
         name = link['href'].split('/')[2][:-4]
         cover_downloaded = inspect_game(name, alt_names)
+        game = cover_downloaded[2]
 
         # If no cover was present in the game page download from game list
         if not cover_downloaded[0]:
@@ -231,13 +255,20 @@ for c in string.ascii_uppercase:
             for img in imgs:
                 if "Images-Minicovers" in img['src']:
                     r = requests.get(f"http://www.pcengine.co.uk/{img['src'][3:]}")
-                    with open(f'output/pce/{cover_downloaded[1]}.jpg', 'wb') as f:
-                        f.write(r.content)
-                    call(
-                        f'convert "output/pce/{cover_downloaded[1]}.jpg" -filter lanczos -resize 64x64! -normalize -unsharp 0 -enhance +dither -colors 64 "output/pce/{cover_downloaded[1]}.png"', shell=True)
-                    optimize_image(f"output/pce/{cover_downloaded[1]}.png")
+                    if not os.path.isfile(f'output/pce/{cover_downloaded[1]}.jpg'):
+                        with open(f'output/pce/{cover_downloaded[1]}.jpg', 'wb') as f:
+                            f.write(r.content)
+                        call(
+                            f'convert "output/pce/{cover_downloaded[1]}.jpg" -filter lanczos -resize 64x64! -normalize -unsharp 0 -enhance +dither "output/pce/{cover_downloaded[1]}.bmp3"', shell=True)
+                        os.rename(f'output/pce/{cover_downloaded[1]}.bmp3', f'output/pce/{cover_downloaded[1]}.bmp')
+                        call(f'python3 Quantomatic.py "output/pce/{cover_downloaded[1]}" -p 13', shell=True)
+                        call(f'convert "output/pce/{cover_downloaded[1]}CrushedPals.bmp" "output/pce/{cover_downloaded[1]}.png"', shell=True)
+                        os.remove(f'output/pce/{cover_downloaded[1]}.bmp')
+                        os.remove(f'output/pce/{cover_downloaded[1]}CrushedPals.bmp')
+                    game['image'] = f'output/pce/{cover_downloaded[1]}.png'
+                    #optimize_image(f"output/pce/{cover_downloaded[1]}.png")
+        db.append(game)
         print(name)
-
 
 with open('db.json', 'w') as f:
     json.dump(db, f, indent=2)
